Update the demo applications to use the updated BIO client API

+++ demo_get.c
@@ -91,7 +91,7 @@ use_high_level_api(int argc, char **argv)
     }
     
     char *key = NULL;
-    size_t key_size = 0;
+    int key_size = 0;
     size_t id_size = kmip_strnlen_s(argv[1], 50);
     
     result = kmip_bio_get_symmetric_key(
@@ -117,7 +117,7 @@ use_high_level_api(int argc, char **argv)
         if(result == KMIP_STATUS_SUCCESS)
         {
             printf("Symmetric Key ID: %s\n", argv[1]);
-            printf("Symmetric Key Size: %zu bits\n", key_size * 8);
+            printf("Symmetric Key Size: %d bits\n", key_size * 8);
             printf("Symmetric Key:");
             print_buffer(key, key_size);
             printf("\n");
@@ -202,7 +202,7 @@ use_mid_level_api(int argc, char **argv)
     }
     
     char *key = NULL;
-    size_t key_size = 0;
+    int key_size = 0;
     size_t id_size = kmip_strnlen_s(argv[1], 50);
     
     KMIP kmip_context = {0};
@@ -237,7 +237,7 @@ use_mid_level_api(int argc, char **argv)
         if(result == KMIP_STATUS_SUCCESS)
         {
             printf("Symmetric Key ID: %s\n", argv[1]);
-            printf("Symmetric Key Size: %zu bits\n", key_size * 8);
+            printf("Symmetric Key Size: %d bits\n", key_size * 8);
             printf("Symmetric Key:");
             print_buffer(key, key_size);
             printf("\n");
